app/requirements.txt
flask-mongoengine
flask_user
flask-bcrypt
flask_jwt_extended